table: renderHint in separate file
